requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
google-api-core==1.22.3
google-api-python-client==1.12.3
google-auth==1.22.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
html5lib==1.0.1
httplib2==0.18.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
multidict==4.7.6
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycryptodome==3.9.9
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
selenium==3.141.0
six==1.14.0
tabulate==0.8.7
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
yarl==1.6.0
